+++ R/cpma.R
@@ -7,20 +7,20 @@
 #'
 #' @param p test
 #' @param epsilon test
-#' @param .THRESHOLD test
-#' @param .MAXVAL test
-#' @param .FIXED test
+#' @param p_threshold test
+#' @param maxval test
+#' @param fixed test
 #' @importFrom stats pchisq
 #' @export
 
 cpma <- function(p,
                  epsilon = 0.001,
-                 .THRESHOLD = 5E-8,
-                 .MAXVAL = 1,
-                 .FIXED = TRUE) {
+                 p_threshold = 5E-8,
+                 maxval = 1,
+                 fixed = TRUE) {
 
   # check input arguments
-  p_args <- check_p(p, .MAXVAL)
+  p_args <- check_p(p, maxval)
   m <- p_args$number_of_traits
   l <- p_args$number_of_snps
   pm <- p_args$p_matrix
@@ -28,10 +28,10 @@ cpma <- function(p,
   if (!is.numeric(epsilon) || epsilon < 0 || epsilon > 1)
     stop("Make sure `epsilon` is a number between 0 and 1")
 
-  if (!is.numeric(.THRESHOLD) || .THRESHOLD < 0 || .THRESHOLD > 1)
-    stop("Make sure `.THRESHOLD` is a number between 0 and 1")
+  if (!is.numeric(p_threshold) || p_threshold < 0 || p_threshold > 1)
+    stop("Make sure `p_threshold` is a number between 0 and 1")
 
-  if (!is.logical(.FIXED)) stop("`.FIXED` should be `TRUE` or `FALSE`")
+  if (!is.logical(fixed)) stop("`fixed` should be `TRUE` or `FALSE`")
 
   # run
   final <- make_final(l)
@@ -46,12 +46,12 @@ cpma <- function(p,
     p2 <- pchisq(abs(oo - ee) * 2, 1, lower.tail = FALSE)
     p3 <- min(p2)
     n <- which.min(p2)
-    n2 <- if (p3 < .THRESHOLD) n else 0
+    n2 <- if (p3 < p_threshold) n else 0
     traits <- o$ix[seq_len(n2)]
     p_out <- p2[n]
 
-    nn <- sum(o$x < .THRESHOLD)
-    if (.FIXED && nn > n2) {
+    nn <- sum(o$x < p_threshold)
+    if (fixed && nn > n2) {
       n2 <- nn
       p_out <- p2[nn]
       traits <- o$ix[1:nn]