+++ tests/testthat/test-cpma.R
@@ -38,13 +38,13 @@ test_that("cpma can deal with illogical values", {
   expect_error(cpma(p_2, epsilon = -1))
   expect_error(cpma(p_2, epsilon = 2))
 
-  expect_error(cpma(p_2, .THRESHOLD = "test"))
-  expect_error(cpma(p_2, .THRESHOLD = -1))
-  expect_error(cpma(p_2, .THRESHOLD = 2))
+  expect_error(cpma(p_2, p_threshold = "test"))
+  expect_error(cpma(p_2, p_threshold = -1))
+  expect_error(cpma(p_2, p_threshold = 2))
 
-  expect_error(cpma(p_2, .MAXVAL = "test"))
-  expect_error(cpma(p_2, .MAXVAL = -1))
-  expect_error(cpma(p_2, .MAXVAL = 2))
+  expect_error(cpma(p_2, maxval = "test"))
+  expect_error(cpma(p_2, maxval = -1))
+  expect_error(cpma(p_2, maxval = 2))
 })
 
 # handling extreme values