+++ tests/testthat/test-fct.R
@@ -34,13 +34,13 @@ test_that("fct can deal with illogical values", {
   expect_error(fct(p_neg))
   expect_error(fct(p_char))
 
-  expect_error(fct(p_2, .THRESHOLD = "test"))
-  expect_error(fct(p_2, .THRESHOLD = -1))
-  expect_error(fct(p_2, .THRESHOLD = 2))
+  expect_error(fct(p_2, p_threshold = "test"))
+  expect_error(fct(p_2, p_threshold = -1))
+  expect_error(fct(p_2, p_threshold = 2))
 
-  expect_error(fct(p_2, .MAXVAL = "test"))
-  expect_error(fct(p_2, .MAXVAL = -1))
-  expect_error(fct(p_2, .MAXVAL = 2))
+  expect_error(fct(p_2, maxval = "test"))
+  expect_error(fct(p_2, maxval = -1))
+  expect_error(fct(p_2, maxval = 2))
 })
 
 # handling extreme values